Move `ReceiptInfo` to the eth-rpc
