+++ substrate/frame/revive/rpc/examples/eth-rpc-tester.rs
@@ -16,8 +16,8 @@
 // limitations under the License.
 use clap::Parser;
 use jsonrpsee::http_client::HttpClientBuilder;
-use pallet_revive::evm::{Account, BlockTag, ReceiptInfo};
-use pallet_revive_eth_rpc::{EthRpcClient, example::TransactionBuilder};
+use pallet_revive::evm::{Account, BlockTag};
+use pallet_revive_eth_rpc::{EthRpcClient, ReceiptInfo, example::TransactionBuilder};
 use std::sync::Arc;
 use tokio::{
 	io::{AsyncBufReadExt, BufReader},
