+++ substrate/frame/revive/rpc/src/lib.rs
@@ -476,7 +476,7 @@ impl EthRpcServer for EthRpcServerImpl {
 		let receipt = self.client.receipt(&transaction_hash).await;
 		let signed_tx = self.client.signed_tx_by_hash(&transaction_hash).await;
 		if let (Some(receipt), Some(signed_tx)) = (receipt, signed_tx) {
-			return Ok(Some(TransactionInfo::new(&receipt, signed_tx)));
+			return Ok(Some(receipt.transaction_info(signed_tx)));
 		}
 
 		Ok(None)
@@ -567,7 +567,7 @@ impl EthRpcServerImpl {
 			return Ok(None);
 		};
 
-		Ok(Some(TransactionInfo::new(&receipt, signed_tx)))
+		Ok(Some(receipt.transaction_info(signed_tx)))
 	}
 
 	async fn handle_subscription_forwarding(
