+++ substrate/frame/revive/rpc/src/types.rs
@@ -23,3 +23,6 @@ pub use subscription::*;
 
 mod sync;
 pub use sync::*;
+
+mod receipt;
+pub use receipt::*;
