+++ substrate/frame/revive/src/evm/api/rpc_types.rs
@@ -16,11 +16,9 @@
 // limitations under the License.
 //! Utility impl for the RPC types.
 use super::*;
-use alloc::vec::Vec;
 use codec::{Decode, Encode};
 use scale_info::TypeInfo;
 use sp_core::{H160, U256};
-use sp_crypto_hashing::keccak_256;
 
 /// Configuration specific to a dry-run execution.
 ///
@@ -228,86 +226,6 @@ impl From<TransactionSigned> for TransactionUnsigned {
 	}
 }
 
-impl TransactionInfo {
-	/// Create a new [`TransactionInfo`] from a receipt and a signed transaction.
-	pub fn new(receipt: &ReceiptInfo, transaction_signed: TransactionSigned) -> Self {
-		Self {
-			block_hash: receipt.block_hash,
-			block_number: receipt.block_number,
-			from: receipt.from,
-			hash: receipt.transaction_hash,
-			transaction_index: receipt.transaction_index,
-			transaction_signed,
-		}
-	}
-}
-
-impl ReceiptInfo {
-	/// Initialize a new Receipt
-	pub fn new(
-		block_hash: H256,
-		block_number: U256,
-		contract_address: Option<Address>,
-		from: Address,
-		logs: Vec<Log>,
-		to: Option<Address>,
-		effective_gas_price: U256,
-		gas_used: U256,
-		success: bool,
-		transaction_hash: H256,
-		transaction_index: U256,
-		r#type: Byte,
-	) -> Self {
-		let logs_bloom = Self::logs_bloom(&logs);
-		ReceiptInfo {
-			block_hash,
-			block_number,
-			contract_address,
-			from,
-			logs,
-			logs_bloom,
-			to,
-			effective_gas_price,
-			gas_used,
-			status: Some(if success { U256::one() } else { U256::zero() }),
-			transaction_hash,
-			transaction_index,
-			r#type: Some(r#type),
-			..Default::default()
-		}
-	}
-
-	/// Returns `true` if the transaction was successful.
-	pub fn is_success(&self) -> bool {
-		self.status.map_or(false, |status| status == U256::one())
-	}
-
-	/// Calculate receipt logs bloom.
-	fn logs_bloom(logs: &[Log]) -> Bytes256 {
-		let mut bloom = [0u8; 256];
-		for log in logs {
-			m3_2048(&mut bloom, &log.address.as_ref());
-			for topic in &log.topics {
-				m3_2048(&mut bloom, topic.as_ref());
-			}
-		}
-		bloom.into()
-	}
-}
-/// Specialised Bloom filter that sets three bits out of 2048, given an
-/// arbitrary byte sequence.
-///
-/// See Section 4.4.1 "Transaction Receipt" of the [Ethereum Yellow Paper][ref].
-///
-/// [ref]: https://ethereum.github.io/yellowpaper/paper.pdf
-fn m3_2048(bloom: &mut [u8; 256], bytes: &[u8]) {
-	let hash = keccak_256(bytes);
-	for i in [0, 2, 4] {
-		let bit = (hash[i + 1] as usize + ((hash[i] as usize) << 8)) & 0x7FF;
-		bloom[256 - 1 - bit / 8] |= 1 << (bit % 8);
-	}
-}
-
 #[test]
 fn can_deserialize_input_or_data_field_from_generic_transaction() {
 	let cases = [
@@ -347,47 +265,6 @@ fn test_block_number_or_tag_or_hash_deserialization() {
 		let val: BlockNumberOrTagOrHash = serde_json::from_str(s).unwrap();
 		assert_eq!(val, BlockNumberOrTagOrHash::BlockHash(H256([0xaau8; 32])));
 	}
-}
-
-#[test]
-fn logs_bloom_works() {
-	let receipt: ReceiptInfo = serde_json::from_str(
-		r#"
-		{
-			"blockHash": "0x835ee379aaabf4802a22a93ad8164c02bbdde2cc03d4552d5c642faf4e09d1f3",
-			"blockNumber": "0x2",
-			"contractAddress": null,
-			"cumulativeGasUsed": "0x5d92",
-			"effectiveGasPrice": "0x2dcd5c2d",
-			"from": "0xb4f1f9ecfe5a28633a27f57300bda217e99b8969",
-			"gasUsed": "0x5d92",
-			"logs": [
-				{
-				"address": "0x82bdb002b9b1f36c42df15fbdc6886abcb2ab31d",
-				"topics": [
-					"0x1585375487296ff2f0370daeec4214074a032b31af827c12622fa9a58c16c7d0",
-					"0x000000000000000000000000b4f1f9ecfe5a28633a27f57300bda217e99b8969"
-				],
-				"data": "0x00000000000000000000000000000000000000000000000000000000000030390000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000b48656c6c6f20776f726c64000000000000000000000000000000000000000000",
-				"blockNumber": "0x2",
-				"transactionHash": "0xad0075127962bdf73d787f2944bdb5f351876f23c35e6a48c1f5b6463a100af4",
-				"transactionIndex": "0x0",
-				"blockHash": "0x835ee379aaabf4802a22a93ad8164c02bbdde2cc03d4552d5c642faf4e09d1f3",
-				"logIndex": "0x0",
-				"removed": false
-				}
-			],
-			"logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000008000000000000000000000000000000000000000000000000800000000040000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000004000000000000000800000000000000000080000000000000000000000000000000000000000000",
-			"status": "0x1",
-			"to": "0x82bdb002b9b1f36c42df15fbdc6886abcb2ab31d",
-			"transactionHash": "0xad0075127962bdf73d787f2944bdb5f351876f23c35e6a48c1f5b6463a100af4",
-			"transactionIndex": "0x0",
-			"type": "0x2"
-		}
-		"#,
-	)
-	.unwrap();
-	assert_eq!(receipt.logs_bloom, ReceiptInfo::logs_bloom(&receipt.logs));
 }
 
 impl GenericTransaction {
